// src/declarations/enums.ts
export enum Language {
  Cpp = "C++",
  C = "C",
}

export enum Optimization {
  Off = "off",
  Level1 = "level1",
  Level2 = "level2",
  Level3 = "level3",
}

export enum ProjectType {
  Executable = "executable",
  Library = "library",
}

export enum LibraryType {
  Static = "static",
  Shared = "shared",
}

export enum TemplateConfiguration {
  Empty = "",
  Debug = "debug",
  Release = "release",
}